% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdeco.pg.xc.R
\name{scdeco.pg.xc}
\alias{scdeco.pg.xc}
\title{ZENCO fitting function when secondary covariate is provided}
\usage{
scdeco.pg.xc(
  dat,
  b0,
  b1,
  adapt_iter = 10000,
  update_iter = 5000,
  coda_iter = 50000,
  coda_thin = 20,
  coda_burnin = 10000
)
}
\arguments{
\item{dat}{matrix containing expression values as first two columns and covariate as third column}

\item{b0}{intercept of zinf parameter}

\item{b1}{slope of zinf parameter}

\item{adapt_iter}{number of adaptation iterations in the jags.model function}

\item{update_iter}{update iterations in the update function}

\item{coda_iter}{number of iterations for the coda.sample function}

\item{coda_thin}{how much to thin the resulting MCMC output}

\item{coda_burnin}{how many iterations to burn before beginning coda sample collection}
}
\value{
MCMC samples that have been adapted, burned, and thinned
}
\description{
ZENCO fitting function when secondary covariate is provided
}
\keyword{internal}
